telecom-mnp-api/requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
pydantic
pydantic-settings
python-dotenv

passlib[bcrypt]
python-jose[cryptography]

redis

python-multipart

pytest
httpx

jwt